small refactor to follow common style
