+++ pom.xml
@@ -4,7 +4,7 @@
   <parent>
     <groupId>io.akka</groupId>
     <artifactId>akka-javasdk-parent</artifactId>
-    <version>3.4.0-M4-18-718e8164-dev-SNAPSHOT</version>
+    <version>3.4.0-M6</version>
   </parent>
   
   <groupId>com.example</groupId>
