+++ src/main/java/realestate/domain/ClientState.java
@@ -1,0 +1,4 @@
+package realestate.domain;
+
+public record ClientState(String name, String email, String phone) {
+}
